--- src/main/java/com/springbootstudy/app/controller/WidgetController.java
@@ -1,90 +1,0 @@
-package com.springbootstudy.app.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.json.JSONObject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Controller;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-import java.io.*;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
-
-@Controller
-public class WidgetController {
-
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-    @RequestMapping(value = "/confirm", method = RequestMethod.POST)
-    public ResponseEntity<JSONObject> confirmPayment(@RequestBody String jsonBody) throws Exception {
-        JSONObject requestData = new JSONObject(jsonBody);
-        String paymentKey = requestData.getString("paymentKey");
-        String orderId = requestData.getString("orderId");
-        String amount = requestData.getString("amount");
-
-        JSONObject obj = new JSONObject();
-        obj.put("orderId", orderId);
-        obj.put("amount", amount);
-        obj.put("paymentKey", paymentKey);
-
-        String widgetSecretKey = "test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6";
-
-        Base64.Encoder encoder = Base64.getEncoder();
-        byte[] encodedBytes = encoder.encode((widgetSecretKey + ":").getBytes(StandardCharsets.UTF_8));
-        String authorizations = "Basic " + new String(encodedBytes);
-
-        URL url = new URL("https://api.tosspayments.com/v1/payments/confirm");
-        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-        connection.setRequestProperty("Authorization", authorizations);
-        connection.setRequestProperty("Content-Type", "application/json");
-        connection.setRequestMethod("POST");
-        connection.setDoOutput(true);
-
-        OutputStream outputStream = connection.getOutputStream();
-        outputStream.write(obj.toString().getBytes("UTF-8"));
-
-        int code = connection.getResponseCode();
-        boolean isSuccess = code == 200;
-
-        InputStream responseStream = isSuccess ? connection.getInputStream() : connection.getErrorStream();
-
-        BufferedReader reader = new BufferedReader(new InputStreamReader(responseStream, StandardCharsets.UTF_8));
-        StringBuilder response = new StringBuilder();
-        String line;
-        while ((line = reader.readLine()) != null) {
-            response.append(line);
-        }
-        reader.close();
-
-        JSONObject jsonObject = new JSONObject(response.toString());
-        return ResponseEntity.status(code).body(jsonObject);
-    }
-
-    @RequestMapping(value = "/success", method = RequestMethod.GET)
-    public String paymentRequest(HttpServletRequest request, Model model) throws Exception {
-        return "/success";
-    }
-
-    @RequestMapping(value = "/", method = RequestMethod.GET)
-    public String index(HttpServletRequest request, Model model) throws Exception {
-        return "/checkout";
-    }
-
-    @RequestMapping(value = "/fail", method = RequestMethod.GET)
-    public String failPayment(HttpServletRequest request, Model model) throws Exception {
-        String failCode = request.getParameter("code");
-        String failMessage = request.getParameter("message");
-
-        model.addAttribute("code", failCode);
-        model.addAttribute("message", failMessage);
-
-        return "/fail";
-    }
-}
