+++ src/main/resources/templates/fragments/headerShop.html
@@ -19,7 +19,7 @@
 				<th:block th:if="${session.isLogin}">
 					<a class="logout2" href="/logout">로그아웃&nbsp;&nbsp;</a>
 					<a class="joinupdate2" href="/joinupdate">|&nbsp;&nbsp;마이페이지&nbsp;&nbsp;</a>
-					<a class="cart2" href="/cart">|&nbsp;&nbsp;<i class="bi bi-cart4"></i> (0)&nbsp;&nbsp;</a>
+					<a class="cart2" href="/shopCart">|&nbsp;&nbsp;<i class="bi bi-cart4"></i> (0)&nbsp;&nbsp;</a>
 				</th:block>
 				<th:block th:unless="${session.isLogin}">
 					<a class="login2" href="/login">로그인&nbsp;&nbsp;</a>
