+++ src/main/resources/templates/views/shop/shopCart.html
@@ -16,7 +16,7 @@
         <div class="col-8" id="productContainer">
             <div th:if="${cartDetails != null}">
                 <!-- 장바구니 품목 -->
-                <th:block th:each="item : ${cartDetails}">
+                <th:block th:each="item : ${cartDetails}" >
                 <div class="row cartBox" th:data-product-id="${item.productId}">
                     <div class="col productImage"><img src="http://via.placeholder.com/140x140"></div>
                     <div class="col">
